refactor: move ResultView method to InputView
