+++ src/main/java/cc/oakk/lotto/view/InputView.java
@@ -5,4 +5,7 @@ import java.util.List;
 public interface InputView {
     int readMoney();
     List<Integer> readWinningNumbers();
+
+    void printMoneyInputHeader();
+    void printWinningNumberInputHeader();
 }
