+++ src/main/java/cc/oakk/lotto/view/ResultView.java
@@ -4,13 +4,9 @@ import cc.oakk.lotto.model.LottoResults;
 import cc.oakk.lotto.view.printer.Printable;
 
 public interface ResultView {
-    void printMoneyInputHeader();
     void repeatPurchasedLottoCount(int count);
     void printLottos(Printable<String> lottosPrintable);
 
-    void printWinningNumberInputHeader();
-
     void printResultHeader();
     void printLottoResults(Printable<LottoResults> lottoResultsPrintable);
-
 }
