+++ src/main/java/cc/oakk/lotto/view/impl/InputConsoleView.java
@@ -1,6 +1,8 @@
 package cc.oakk.lotto.view.impl;
 
 import cc.oakk.lotto.view.InputView;
+import cc.oakk.lotto.view.printer.Printer;
+import cc.oakk.lotto.view.printer.StandardPrinter;
 
 import java.util.Arrays;
 import java.util.List;
@@ -10,13 +12,16 @@ import java.util.stream.Collectors;
 public class InputConsoleView implements InputView {
     private static final String NUMBER_DELIMITER = ", ";
     private final Scanner scanner;
+    private final Printer<String> stringPrinter;
 
     public InputConsoleView() {
         this.scanner = new Scanner(System.in);
+        this.stringPrinter = StandardPrinter.getInstance();
     }
 
-    public InputConsoleView(Scanner scanner) {
+    public InputConsoleView(Scanner scanner, Printer<String> stringPrinter) {
         this.scanner = scanner;
+        this.stringPrinter = stringPrinter;
     }
 
     @Override
@@ -33,5 +38,15 @@ public class InputConsoleView implements InputView {
         return Arrays.stream(input.split(NUMBER_DELIMITER))
                 .map(Integer::parseInt)
                 .collect(Collectors.toList());
+    }
+
+    @Override
+    public void printMoneyInputHeader() {
+        stringPrinter.print("구입금액을 입력해 주세요."::toString);
+    }
+
+    @Override
+    public void printWinningNumberInputHeader() {
+        stringPrinter.print("\n지난 주 당첨 번호를 입력해 주세요."::toString);
     }
 }
